Move util methods to MCPTransportBase.
